refactor: modularized MyAccount page into separate components
