+++ front/src/pages/MyAccount/index.tsx
@@ -1,406 +1,33 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { FC, useState, useEffect } from "react";
-import {
-  FaUserCircle,
-  FaEdit,
-  FaSignOutAlt,
-  FaCopy,
-  FaUserPlus,
-  FaCheck,
-  FaTimes,
-} from "react-icons/fa";
-import { useNavigate } from "react-router-dom";
-
+import { FC, useState } from "react";
 import { useAuth } from "../../context/authContext";
-import Swal from "sweetalert2";
 
-import { handleSearch } from "../../services/searchUserService";
-import { sendFriendRequest } from "../../services/sendFriendRequest";
-import {
-  fetchFriendRequests,
-  acceptFriendRequest,
-  declineFriendRequest,
-} from "../../services/friendRequest";
-import { fetchContacts } from "../../services/fetchContacts";
+import ContactList from "./ContactList";
+import FriendRequests from "./FriendRequests";
+import AccountHeader from "./AccountHeader";
+import AddContactModal from "./AddContactModal";
 
 const MyAccount: FC = () => {
-  const [contacts, setContacts] = useState<any[]>([]);
-
-  const [searchTerm, setSearchTerm] = useState("");
-
-  // Contatos filtrados
-  const filteredContacts = contacts.filter((contact) => {
-    const name = contact.name || contact.displayName || "";
-    const code = contact.userCode || "";
-
-    return (
-      name.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      code.toLowerCase().includes(searchTerm.toLowerCase())
-    );
-  });
-
   const [showModal, setShowModal] = useState(false);
-  const [searchUser, setSearchUser] = useState("");
+  
+  const { currentUser} = useAuth();
 
-  const [friendRequests, setFriendRequests] = useState<any[]>([]);
-
-  const [results, setResults] = useState<any[]>([]);
-  const { currentUser, logOut: firebaseLogout } = useAuth();
-  const [copied, setCopied] = useState(false);
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    const fetchRequests = async () => {
-      if (currentUser?.uid) {
-        const requests = await fetchFriendRequests(currentUser.uid);
-        setFriendRequests(requests);
-      }
-    };
-
-    fetchRequests();
-
-    const loadContacts = async () => {
-      if (currentUser?.uid) {
-        const contactList = await fetchContacts();
-        setContacts(contactList);
-
-        console.log("Contatos carregados:", contactList);
-      }
-    };
-
-    loadContacts();
-  }, [currentUser?.uid]);
-
-  const onSearchInputChange = async (value: string) => {
-    setSearchUser(value);
-    const res = await handleSearch(value);
-    setResults(res);
-  };
-
-  const handleCopy = () => {
-    navigator.clipboard.writeText(currentUser?.userCode || "");
-    setCopied(true);
-    setTimeout(() => setCopied(false), 2000);
-  };
-
-  const handleLogout = async () => {
-    try {
-      await firebaseLogout();
-      Swal.fire({
-        icon: "success",
-        title: "Logout realizado!",
-        text: "Você foi desconectado com sucesso.",
-        timer: 1500,
-        showConfirmButton: false,
-        background: "#1E293B",
-        color: "#E0E7FF",
-      });
-      navigate("/login-register");
-    } catch (error) {
-      console.error("Erro ao fazer logout:", error);
-      Swal.fire({
-        icon: "error",
-        title: "Erro",
-        text: "Não foi possível fazer logout.",
-        background: "#1E293B",
-        color: "#E0E7FF",
-      });
-    }
-  };
 
   return (
     <section className="py-10 mt-24 flex justify-center items-center min-h-screen bg-gradient-to-br from-[#0F172A] to-[#1E293B] px-4">
-      <div className="bg-[#1E293B] border border-blue-800 text-center text-blue-100 p-8 rounded-2xl shadow-xl w-full max-w-md">
-        {/* Avatar */}
-        <div className="flex justify-center mb-4">
-          {currentUser?.photoURL ? (
-            <img
-              src={currentUser.photoURL}
-              alt="Avatar"
-              className="w-24 h-24 rounded-full border-4 border-blue-700 shadow-lg"
-            />
-          ) : (
-            <FaUserCircle className="text-blue-500" size={80} />
-          )}
-        </div>
-
-        {/* Nome e email */}
-        <h2 className="text-2xl font-semibold mb-1">
-          {currentUser?.displayName}
-        </h2>
-        <p className="text-blue-300 text-sm mb-4">{currentUser?.email}</p>
-
-        {/* Código do usuário */}
-        <div className="bg-[#0F172A] border border-blue-700 rounded-lg p-3 mb-6 flex items-center justify-between">
-          <span className="text-sm font-mono text-blue-300">
-            {currentUser?.userCode}
-          </span>
-          <button
-            onClick={handleCopy}
-            className="text-blue-400 cursor-pointer hover:text-blue-200 transition"
-            title="Copiar código"
-          >
-            <FaCopy />
-          </button>
-        </div>
-        {copied && (
-          <div className="flex items-center justify-center gap-2 text-green-400 bg-green-900/30 border border-green-700 px-4 py-2 rounded-lg text-sm mb-6 shadow-md transition-opacity">
-            <svg
-              className="w-4 h-4 text-green-400"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M5 13l4 4L19 7"
-              />
-            </svg>
-            Código copiado com sucesso!
-          </div>
-        )}
-
-        {/* Ações */}
-        <div className="flex flex-col gap-4 mb-8">
-          <button className="flex items-center justify-center cursor-pointer gap-2 bg-[#3B82F6] hover:bg-[#1E40AF] text-white font-medium py-2 px-4 rounded-lg transition-all">
-            <FaEdit /> Editar Perfil
-          </button>
-
-          <button
-            onClick={() => setShowModal(true)}
-            className="flex items-center justify-center cursor-pointer gap-2 bg-[#3B82F6] hover:bg-[#1E40AF] text-white font-medium py-2 px-4 rounded-lg transition-all"
-          >
-            <FaUserPlus /> Adicionar Contato
-          </button>
-
-          <button
-            onClick={handleLogout}
-            className="flex items-center justify-center cursor-pointer gap-2 bg-red-500 hover:bg-red-600 text-white font-medium py-2 px-4 rounded-lg transition-all"
-          >
-            <FaSignOutAlt /> Sair da Conta
-          </button>
-        </div>
+      <div className="bg-[rgb(30,41,59)] border border-blue-800 text-center text-blue-100 p-8 rounded-2xl shadow-xl w-full max-w-md">
+        { /* Cabeçalho da Conta */}
+        <AccountHeader currentUser={currentUser} setShowModal={setShowModal} />
 
         {/* Contatos */}
-        <div className="text-left mt-6">
-          <h3 className="text-lg font-semibold text-blue-300 mb-3 ml-2">
-            Seus Contatos
-          </h3>
-
-          {/* Barra de busca */}
-          <div className="relative mb-4 flex justify-center items-center">
-            <span className="absolute left-4 top-1/2 transform -translate-y-1/2 text-blue-400">
-              <svg
-                className="w-5 h-5"
-                fill="none"
-                stroke="currentColor"
-                strokeWidth="2"
-                viewBox="0 0 24 24"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  d="M21 21l-4.35-4.35m0 0A7.5 7.5 0 1010.5 18a7.5 7.5 0 006.15-3.35z"
-                />
-              </svg>
-            </span>
-            <input
-              type="text"
-              placeholder="Buscar por nome ou código..."
-              value={searchTerm}
-              onChange={(e) => setSearchTerm(e.target.value)}
-              className="w-[96%] pl-10 py-2 rounded-lg bg-[#0F172A] border border-blue-700 text-blue-100 placeholder:text-blue-400 focus:outline-none focus:ring-2 focus:ring-blue-500"
-            />
-          </div>
-
-          <ul className="space-y-2 max-h-52 overflow-y-auto custom-scrollbar px-2">
-            {filteredContacts.length > 0 ? (
-              filteredContacts.map((contact, index) => (
-                <li
-                  key={index}
-                  className="p-3 rounded-lg bg-[#0F172A] border border-blue-700 hover:bg-blue-800/30 transition"
-                >
-                  <p className="font-medium text-blue-100">
-                    {contact.displayName}
-                  </p>
-                  <p className="text-sm text-blue-400">{contact.email}</p>
-                </li>
-              ))
-            ) : (
-              <li className="text-sm text-blue-400">
-                Nenhum contato encontrado.
-              </li>
-            )}
-          </ul>
-        </div>
+        <ContactList currentUser={currentUser} />
 
         {/* Pedidos de Amizade Recebidos */}
-        <div className="text-left mt-10">
-          <h3 className="text-lg font-semibold text-blue-300 mb-3 ml-2">
-            Pedidos de Amizade
-          </h3>
-
-          {friendRequests.length > 0 ? (
-            <ul className="space-y-2 max-h-52 overflow-y-auto custom-scrollbar px-2">
-              {friendRequests.map((req) => (
-                <li
-                  key={req.requesterId}
-                  className="p-3 rounded-lg bg-[#0F172A] border border-blue-700 flex justify-between items-center hover:bg-blue-800/30 transition"
-                >
-                  <div>
-                    <p className="font-medium text-blue-100">
-                      {req.requesterDisplayName || req.requesterId}
-                    </p>
-                    <p className="text-sm text-blue-400">
-                      Status: {req.status}
-                    </p>
-                  </div>
-                  <div className="flex gap-2">
-                    <button
-                      onClick={async () => {
-                        if (currentUser?.uid) {
-                          await acceptFriendRequest(
-                            req.requesterId,
-                            currentUser.uid
-                          );
-                          setFriendRequests((prev) =>
-                            prev.filter(
-                              (r) => r.requesterId !== req.requesterId
-                            )
-                          );
-                        }
-                      }}
-                      className="bg-green-600 hover:bg-green-700 text-white px-2 cursor-pointer py-1 rounded text-lg flex items-center justify-center"
-                      title="Aceitar"
-                    >
-                      <FaCheck />
-                    </button>
-
-                    <button
-                      onClick={async () => {
-                        if (currentUser?.uid) {
-                          await declineFriendRequest(
-                            req.requesterId,
-                            currentUser.uid
-                          );
-                          setFriendRequests((prev) =>
-                            prev.filter(
-                              (r) => r.requesterId !== req.requesterId
-                            )
-                          );
-                        }
-                      }}
-                      className="bg-red-500 hover:bg-red-600 text-white px-2 py-1 cursor-pointer rounded text-lg flex items-center justify-center"
-                      title="Recusar"
-                    >
-                      <FaTimes />
-                    </button>
-                  </div>
-                </li>
-              ))}
-            </ul>
-          ) : (
-            <p className="text-sm text-blue-400 ml-2">
-              Nenhum pedido recebido.
-            </p>
-          )}
-        </div>
+        <FriendRequests currentUser={currentUser} />
       </div>
 
       {showModal && (
-        <div
-          className="fixed inset-0 bg-black/70 flex items-center justify-center z-50 backdrop-blur-sm transition-opacity duration-300"
-          onClick={() => setShowModal(false)}
-        >
-          <div
-            className="bg-[#1E293B] text-blue-100 w-full max-w-md rounded-2xl p-6 shadow-2xl border border-blue-700 relative transform transition-all duration-300 scale-95 animate-fadeIn"
-            onClick={(e) => e.stopPropagation()} // <- impede o clique de propagar para o fundo
-          >
-            <button
-              onClick={() => setShowModal(false)}
-              className="absolute top-3 right-4 text-blue-400 hover:text-red-400 text-2xl font-bold transition-transform duration-200 hover:scale-110"
-              aria-label="Fechar"
-            >
-              &times;
-            </button>
-
-            <h2 className="text-xl font-semibold mb-5 text-center tracking-wide">
-              Adicionar Contato
-            </h2>
-
-            {/* Campo de busca com ícone */}
-            <div className="relative mb-5">
-              <span className="absolute inset-y-0 left-3 flex items-center pointer-events-none">
-                <svg
-                  className="w-5 h-5 text-blue-400"
-                  fill="none"
-                  stroke="currentColor"
-                  strokeWidth="2"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M21 21l-4.35-4.35m0 0A7.5 7.5 0 1010.5 18a7.5 7.5 0 006.15-3.35z"
-                  />
-                </svg>
-              </span>
-              <input
-                type="text"
-                value={searchUser}
-                onChange={(e) => onSearchInputChange(e.target.value)}
-                placeholder="Buscar por código ou e-mail"
-                className="w-full pl-10 pr-4 py-2 rounded-lg bg-[#0F172A] border border-blue-600 placeholder:text-blue-400 text-blue-100 focus:outline-none focus:ring-2 focus:ring-blue-500 transition duration-200"
-              />
-            </div>
-
-            {/* Resultados */}
-            <ul className="space-y-3 max-h-52 overflow-y-auto custom-scrollbar">
-              {results.length > 0 ? (
-                results.map((user) => (
-                  <li
-                    key={user.uid}
-                    className="bg-[#0F172A] border border-blue-700 p-3 rounded-lg flex justify-between items-center hover:bg-blue-800/30 transition-colors duration-200"
-                  >
-                    <div>
-                      <p className="font-medium">{user.displayName}</p>
-                      <p className="text-sm text-blue-400">{user.email}</p>
-                    </div>
-                    <button
-                      onClick={async () => {
-                        if (currentUser?.uid && user.uid !== currentUser.uid) {
-                          const success = await sendFriendRequest(
-                            currentUser.uid,
-                            user.uid,
-                            currentUser.displayName || "Usuário"
-                          );
-                          if (success) {
-                            Swal.fire({
-                              icon: "success",
-                              title: "Pedido enviado!",
-                              text: `Pedido de amizade enviado para ${user.displayName}.`,
-                              background: "#1E293B",
-                              color: "#E0E7FF",
-                            });
-                          }
-                        }
-                      }}
-                      className="text-sm bg-[#3B82F6] hover:bg-[#2563EB] text-white px-4 py-1.5 rounded-md transition-all duration-200 cursor-pointer"
-                    >
-                      Enviar pedido
-                    </button>
-                  </li>
-                ))
-              ) : (
-                <p className="text-blue-400 text-sm text-center">
-                  Nenhum usuário encontrado.
-                </p>
-              )}
-            </ul>
-          </div>
-        </div>
+        <AddContactModal currentUser={currentUser} setShowModal={setShowModal} />
       )}
     </section>
   );
